backend, форматирование всего кода с помощью black -S -l 120

+++ backend/tests/test_contacts.py
@@ -10,16 +10,10 @@ class TestContact(BaseTest):
     def test_success_add_contact(self, client: TestClient, db_facade: MockDBFacade):
         new_contact_username = "user4"
         response = client.post(
-            self.contacts_url,
-            headers=self.get_authorization_headers(),
-            json={"login": new_contact_username}
+            self.contacts_url, headers=self.get_authorization_headers(), json={"login": new_contact_username}
         )
         assert response.status_code == 201
-        assert response.json() == {
-            "login": new_contact_username,
-            "name": "user4",
-            "surname": "surname4"
-        }
+        assert response.json() == {"login": new_contact_username, "name": "user4", "surname": "surname4"}
 
         our_user = db_facade.find_user_by_login(login=self.DEFAULT_USER)
         new_contact_user = db_facade.find_user_by_login(login=new_contact_username)
@@ -32,78 +26,47 @@ class TestContact(BaseTest):
 
     def test_add_wrong_token(self, client: TestClient):
         response = client.post(
-            self.contacts_url,
-            headers=self.get_authorization_headers(access_token="cool_token"),
-            json={"login": "user"}
+            self.contacts_url, headers=self.get_authorization_headers(access_token="cool_token"), json={"login": "user"}
         )
         assert response.status_code == 401
         assert response.json() == self.BAD_TOKEN_RESPONSE
 
     def test_add_myself_to_contact(self, client: TestClient):
-        response = client.post(
-            self.contacts_url,
-            headers=self.get_authorization_headers(),
-            json={"login": "user"}
-        )
+        response = client.post(self.contacts_url, headers=self.get_authorization_headers(), json={"login": "user"})
         assert response.status_code == 400
         assert response.json() == self.exception_response("Нельзя добавить себя в контакты")
 
     def test_add_not_exist_user(self, client: TestClient):
         not_exist_user = "not_exist_user"
         response = client.post(
-            self.contacts_url,
-            headers=self.get_authorization_headers(),
-            json={"login": not_exist_user}
+            self.contacts_url, headers=self.get_authorization_headers(), json={"login": not_exist_user}
         )
         assert response.status_code == 404
         assert response.json() == self.exception_response(f"Пользователь с логином '{not_exist_user}' не найден")
 
     def test_add_exist_contact(self, client: TestClient):
-        response = client.post(
-            self.contacts_url,
-            headers=self.get_authorization_headers(),
-            json={"login": "user1"}
-        )
+        response = client.post(self.contacts_url, headers=self.get_authorization_headers(), json={"login": "user1"})
         assert response.status_code == 409
         assert response.json() == self.exception_response("Такой контакт уже существует")
 
     def test_get_many_success(self, client: TestClient):
-        response = client.get(
-            self.contacts_url,
-            headers=self.get_authorization_headers()
-        )
+        response = client.get(self.contacts_url, headers=self.get_authorization_headers())
         assert response.status_code == 200
         assert response.json() == [
-            {
-                "login": "user1",
-                "name": "user1",
-                "surname": "surname1"
-            },
-            {
-                "login": "user2",
-                "name": "user2",
-                "surname": "surname2"
-            },
-            {
-                "login": "user3",
-                "name": "user3",
-                "surname": "surname3"
-            },
+            {"login": "user1", "name": "user1", "surname": "surname1"},
+            {"login": "user2", "name": "user2", "surname": "surname2"},
+            {"login": "user3", "name": "user3", "surname": "surname3"},
         ]
 
     def test_get_many_without_auth(self, client: TestClient):
-        response = client.get(
-            self.contacts_url
-        )
+        response = client.get(self.contacts_url)
         assert response.status_code == 401
         assert response.json() == self.NOT_AUTH_RESPONSE
 
     def test_delete_success(self, client: TestClient, db_facade: MockDBFacade):
         deleted_user_login = "user3"
         response = client.delete(
-            self.contacts_url,
-            headers=self.get_authorization_headers(),
-            json={"login": deleted_user_login}
+            self.contacts_url, headers=self.get_authorization_headers(), json={"login": deleted_user_login}
         )
         assert response.status_code == 200
         assert response.json() == {"message": f"Контакт {deleted_user_login} удалён"}
@@ -118,34 +81,23 @@ class TestContact(BaseTest):
         assert deleted_contact is None
 
     def test_delete_not_auth(self, client: TestClient):
-        response = client.delete(
-            self.contacts_url
-        )
+        response = client.delete(self.contacts_url)
         assert response.status_code == 401
         assert response.json() == self.NOT_AUTH_RESPONSE
 
     def test_delete_miss_contact(self, client: TestClient):
-        response = client.delete(
-            self.contacts_url,
-            headers=self.get_authorization_headers()
-        )
+        response = client.delete(self.contacts_url, headers=self.get_authorization_headers())
         assert response.status_code == 422
 
     def test_delete_not_exist_contact(self, client: TestClient):
-        response = client.delete(
-            self.contacts_url,
-            headers=self.get_authorization_headers(),
-            json={"login": "user5"}
-        )
+        response = client.delete(self.contacts_url, headers=self.get_authorization_headers(), json={"login": "user5"})
         assert response.status_code == 404
         assert response.json() == self.exception_response("Контакт с логином 'user5' не найден")
 
     def test_delete_not_exist_user(self, client: TestClient):
         not_exist_user = "not_exist_user_login"
         response = client.delete(
-            self.contacts_url,
-            headers=self.get_authorization_headers(),
-            json={"login": not_exist_user}
+            self.contacts_url, headers=self.get_authorization_headers(), json={"login": not_exist_user}
         )
         assert response.status_code == 404
         assert response.json() == self.exception_response(f"Пользователь с логином '{not_exist_user}' не найден")
@@ -156,24 +108,15 @@ class TestContact(BaseTest):
             headers=self.get_authorization_headers(),
         )
         assert response.status_code == 200
-        assert response.json() == {
-            "login": "user1",
-            "name": "user1",
-            "surname": "surname1"
-        }
+        assert response.json() == {"login": "user1", "name": "user1", "surname": "surname1"}
 
     def test_get_one_not_auth(self, client: TestClient):
-        response = client.get(
-            f"{self.contacts_url}user_1"
-        )
+        response = client.get(f"{self.contacts_url}user_1")
         assert response.status_code == 401
         assert response.json() == self.NOT_AUTH_RESPONSE
 
     def test_get_one_not_exist_contact(self, client: TestClient):
-        response = client.get(
-            f"{self.contacts_url}user5",
-            headers=self.get_authorization_headers()
-        )
+        response = client.get(f"{self.contacts_url}user5", headers=self.get_authorization_headers())
         assert response.status_code == 404
         assert response.json() == self.exception_response("Контакт с логином 'user5' не найден")
 
@@ -190,11 +133,7 @@ class TestContact(BaseTest):
         response = client.put(
             self.contacts_url,
             headers=self.get_authorization_headers(),
-            json={
-                "login": changed_contact_user,
-                "name": "Пользователь",
-                "surname": "Первый"
-            }
+            json={"login": changed_contact_user, "name": "Пользователь", "surname": "Первый"},
         )
         assert response.status_code == 200
         assert response.json() == {"message": f"Контакт user1 изменён"}
@@ -213,20 +152,13 @@ class TestContact(BaseTest):
         assert changed_contact.contact_user_id == contact_user.id
 
     def test_change_not_auth(self, client: TestClient):
-        response = client.put(
-            self.contacts_url
-        )
+        response = client.put(self.contacts_url)
         assert response.status_code == 401
         assert response.json() == self.NOT_AUTH_RESPONSE
 
     def test_change_bad_data(self, client: TestClient):
         response = client.put(
-            self.contacts_url,
-            headers=self.get_authorization_headers(),
-            json={
-                "name": "name",
-                "surname": "surname"
-            }
+            self.contacts_url, headers=self.get_authorization_headers(), json={"name": "name", "surname": "surname"}
         )
         assert response.status_code == 422
 
@@ -234,11 +166,7 @@ class TestContact(BaseTest):
         response = client.put(
             self.contacts_url,
             headers=self.get_authorization_headers(),
-            json={
-                "login": "user5",
-                "name": "name",
-                "surname": "surname"
-            }
+            json={"login": "user5", "name": "name", "surname": "surname"},
         )
         assert response.status_code == 404
         assert response.json() == self.exception_response("Контакт с логином 'user5' не найден")
@@ -247,11 +175,7 @@ class TestContact(BaseTest):
         response = client.put(
             self.contacts_url,
             headers=self.get_authorization_headers(),
-            json={
-                "login": "not_exist_user_login",
-                "name": "name",
-                "surname": "surname"
-            }
+            json={"login": "not_exist_user_login", "name": "name", "surname": "surname"},
         )
         assert response.status_code == 404
         assert response.json() == self.exception_response("Пользователь с логином 'not_exist_user_login' не найден")
